todo insert api 및 todo log 입력 로직추가

+++ server/repo/todoRepo.js
@@ -1,4 +1,4 @@
-const query = require('../db');
+const { query }= require('../db')
 
 function getTodos(){
     return query(`
@@ -13,17 +13,30 @@ function getTodosByTodoSectionId(todoSectionId){
 
 function insertTodo(title, contents, todoSectionId) {
     return query(`
-        INSERT INTO TodoCard(title, contents, todoSectionId) VALUES ('${title}', '${contents}', ${parseInt(todoSectionId)}));
+        INSERT INTO 
+            TodoCard(title, contents, todoSectionId)
+        VALUES
+            ("${title}", "${contents}", ${parseInt(todoSectionId)});
     `)
 }
 
-function updateTodo(title, contents, id) {
+function updateTodo( id ,title, contents , todoSectionId) {
+    if( !id || ( !todoCardIds && !title && !todoSectionId)) return;
     return query(`
-        UPDATE TodoCard SET title='${title}', contents="${contents}" WHERE id=${parseInt(id)};
+    UPDATE
+        TodoCard
+    SET
+        ${ title ? `title = "${title}"` : ''} 
+        ${ contents ? `contents = "${contents}"` : ''}
+        ${ todoSectionId ? `todoSectionId = ${todoSectionId}` : ''}
+    }    
+    WHERE
+        id = ${parseInt(id)}
     `)
 }
 
 function deleteTodo(id) {
+    if(!id) return;
     return query(`
         DELETE FROM TodoCard WHERE id = ${id};
     `)
